AirplanesLiveHA/requirements.txt
requests>=2.25.1
paho-mqtt>=1.6.1
flask>=2.0.1
pyyaml>=5.4.1 